moved nextgen_main code to process\filter
